refactor common and jwt
